+++ packages/docs/src/pages/SidebarPage.tsx
@@ -20,10 +20,12 @@ export function SidebarPage() {
             <Text>title</Text>
             <Text>title 2</Text>
           </MenuHeader>
-          <MenuItem value="item-1" icon={<House />}>
+          <MenuItem value="item-1" icon={<House />} tooltip="item-1">
             item 1
           </MenuItem>
-          <MenuItem value="item-2">item 2</MenuItem>
+          <MenuItem value="item-2" tooltip="item-2">
+            item 2
+          </MenuItem>
         </Menu>
 
         <SubMenu value="item-1">